requirements.txt
absl-py==2.2.2
aioice==0.9.0
aiortc==1.11.0
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
av==14.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
captcha==0.7.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cors==1.0.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dnspython==2.7.0
extra-streamlit-components==0.1.80
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
future==1.0.0
gast==0.6.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.2
keras-models==0.0.7
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
mplfinance==0.12.10b0
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.12
namex==0.0.8
narwhals==1.34.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
openai-whisper==20240930
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
srsly==2.5.1
streamlit==1.44.1
streamlit-authenticator==0.4.2
streamlit-webrtc==0.61.0
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.2.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yfinance==0.2.55
zope.event==5.0
zope.interface==7.2
